+++ src/actions/treeProperties.js
@@ -1,15 +1,11 @@
-import { calcTipVisibility, calcTipRadii, calcNodeColor } from "../util/treeHelpers";
-import * as controlTypes from "./controls";
-
-export const UPDATE_TIP_VISIBILITY = "UPDATE_TIP_VISIBILITY";
-export const UPDATE_TIP_RADII = "UPDATE_TIP_RADII";
-export const UPDATE_NODE_COLORS = "UPDATE_NODE_COLORS";
+import { calcTipVisibility, calcTipRadii } from "../util/treeHelpers";
+import * as types from "./types";
 
 const updateTipVisibility = () => {
   return (dispatch, getState) => {
     const { tree, metadata, controls } = getState();
     dispatch({
-      type: UPDATE_TIP_VISIBILITY,
+      type: types.UPDATE_TIP_VISIBILITY,
       data: calcTipVisibility(tree, metadata, controls.dateMin, controls.dateMax),
       version: tree.tipVisibilityVersion + 1
     });
@@ -21,10 +17,10 @@ with the new value(s), (b) update the tree tipVisibility */
 export const changeDateFilter = function (newMin, newMax) {
   return (dispatch) => {
     if (newMin) {
-      dispatch({type: controlTypes.CHANGE_DATE_MIN, data: newMin});
+      dispatch({type: types.CHANGE_DATE_MIN, data: newMin});
     }
     if (newMax) {
-      dispatch({type: controlTypes.CHANGE_DATE_MAX, data: newMax});
+      dispatch({type: types.CHANGE_DATE_MAX, data: newMax});
     }
     dispatch(updateTipVisibility());
   };
@@ -35,7 +31,7 @@ const updateTipRadii = () => {
   return (dispatch, getState) => {
     const { controls, sequences, tree } = getState();
     dispatch({
-      type: UPDATE_TIP_RADII,
+      type: types.UPDATE_TIP_RADII,
       data: calcTipRadii(controls.selectedLegendItem, controls.colorScale, sequences, tree),
       version: tree.tipRadiiVersion + 1
     });
@@ -49,25 +45,11 @@ const updateTipRadii = () => {
 export const legendMouseEnterExit = function (label = null) {
   return (dispatch) => {
     if (label) {
-      dispatch({type: controlTypes.LEGEND_ITEM_MOUSEENTER,
+      dispatch({type: types.LEGEND_ITEM_MOUSEENTER,
                 data: label});
     } else {
-      dispatch({type: controlTypes.LEGEND_ITEM_MOUSELEAVE});
+      dispatch({type: types.LEGEND_ITEM_MOUSELEAVE});
     }
     dispatch(updateTipRadii());
-  };
-};
-
-export const updateNodeColors = () => {
-  return (dispatch, getState) => {
-    const { controls, sequences, tree } = getState();
-    const data = calcNodeColor(tree, controls.colorScale, sequences);
-    if (data) {
-      dispatch({
-        type: UPDATE_NODE_COLORS,
-        data: calcNodeColor(tree, controls.colorScale, sequences),
-        version: tree.nodeColorsVersion + 1
-      });
-    }
   };
 };
