+++ src/actions/types.js
@@ -1,0 +1,37 @@
+export const UPDATE_TIP_VISIBILITY = "UPDATE_TIP_VISIBILITY";
+export const UPDATE_TIP_RADII = "UPDATE_TIP_RADII";
+export const UPDATE_NODE_COLORS = "UPDATE_NODE_COLORS";
+export const REQUEST_METADATA = "REQUEST_METADATA";
+export const RECEIVE_METADATA = "RECEIVE_METADATA";
+export const METADATA_FETCH_ERROR = "METADATA_FETCH_ERROR";
+export const REQUEST_TREE = "REQUEST_TREE";
+export const RECEIVE_TREE = "RECEIVE_TREE";
+export const TREE_FETCH_ERROR = "TREE_FETCH_ERROR";
+export const REQUEST_SEQUENCES = "REQUEST_SEQUENCES";
+export const RECEIVE_SEQUENCES = "RECEIVE_SEQUENCES";
+export const SEQUENCES_FETCH_ERROR = "SEQUENCES_FETCH_ERROR";
+export const REQUEST_FREQUENCIES = "REQUEST_FREQUENCIES";
+export const RECEIVE_FREQUENCIES = "RECEIVE_FREQUENCIES";
+export const FREQUENCIES_FETCH_ERROR = "FREQUENCIES_FETCH_ERROR";
+export const REQUEST_ENTROPY = "REQUEST_ENTROPY";
+export const RECEIVE_ENTROPY = "RECEIVE_ENTROPY";
+export const ENTROPY_FETCH_ERROR = "ENTROPY_FETCH_ERROR";
+export const BROWSER_DIMENSIONS = "BROWSER_DIMENSIONS";
+export const TOGGLE_BRANCH_LABELS = "TOGGLE_BRANCH_LABELS";
+export const LEGEND_ITEM_MOUSEENTER = "LEGEND_ITEM_MOUSEENTER";
+export const LEGEND_ITEM_MOUSELEAVE = "LEGEND_ITEM_MOUSELEAVE";
+export const BRANCH_MOUSEENTER = "BRANCH_MOUSEENTER";
+export const BRANCH_MOUSELEAVE = "BRANCH_MOUSELEAVE";
+export const NODE_MOUSEENTER = "NODE_MOUSEENTER";
+export const NODE_MOUSELEAVE = "NODE_MOUSELEAVE";
+export const SEARCH_INPUT_CHANGE = "SEARCH_INPUT_CHANGE";
+export const CHANGE_LAYOUT = "CHANGE_LAYOUT";
+export const CHANGE_DISTANCE_MEASURE = "CHANGE_DISTANCE_MEASURE";
+export const CHANGE_DATE_MIN = "CHANGE_DATE_MIN";
+export const CHANGE_DATE_MAX = "CHANGE_DATE_MAX";
+export const CHANGE_ABSOLUTE_DATE_MIN = "CHANGE_ABSOLUTE_DATE_MIN";
+export const CHANGE_ABSOLUTE_DATE_MAX = "CHANGE_ABSOLUTE_DATE_MAX";
+export const CHANGE_COLOR_BY = "CHANGE_COLOR_BY";
+export const SET_COLOR_SCALE = "SET_COLOR_SCALE";
+export const NEW_DATASET = "NEW_DATASET";
+export const RESET_CONTROLS = "RESET_CONTROLS";
